[AURON #1409] Introduce Spark UI for auron (#1410)
# Which issue does this PR close?

Closes #1409

 # Rationale for this change

Display Auron's build information on the Spark UI.

# What changes are included in this PR?

1. Add a parameter to enable or disable this feature: spark.auron.ui.enabled.
2. Record and parse Auron's build information.
3. Render the Spark UI page.
4. When initializing the Spark session, add the new page through the UI settings.

# Are there any user-facing changes?

Add a new spark.auron.ui.enabled parameter.

# How was this patch tested?

**1. the Spark application is running** 
When the Spark application is running, the Spark UI will display AURON SQL / DataFrame, showing Auron’s build information. 
The information displayed in the Spark UI is as follows:
<img width="1485" height="665" alt="Image" src="https://github.com/user-attachments/assets/1cf8dcab-b42b-45b6-b3a1-4c192ecacafe" />

**2. after the Spark application ends**
After the Spark application ends, the information can be viewed through the Spark History Server. 
You only need to package the auron-spark-ui project into a JAR and integrate it with the Spark History Server, so that the information can be viewed through the Spark History Server after the job finishes.
eg:
auron-spark-ui_2.12-7.0.0-SNAPSHOT.jar to shs jars 

The information displayed in the Spark UI from Spark history server is as follows:

<img width="1316" height="790" alt="Image" src="https://github.com/user-attachments/assets/69ca3cd7-e0ad-4f3e-ac41-b995405b96d7" />

---------

Co-authored-by: guihuawen <guihuawen@didiglobal.com>